chore: update golangci-lint; enable revive & address lints
